+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Display.
 * TokenKind has been removed. 
 * Simplified how tokens are stored in a system. Only the code is now stored.
+* The TokenStore implementation for HashMap has been changed to HashSet, and no longer needs arcs.
 
 ### Removed